Use DistibutionPattern
