+++ compiler/src/iree/compiler/Codegen/Common/GPU/BUILD.bazel
@@ -51,6 +51,7 @@ iree_compiler_cc_library(
         "GPUCheckResourceUsage.cpp",
         "GPUDistribute.cpp",
         "GPUDistributeSharedMemoryCopy.cpp",
+        "GPUDistributionPatterns.cpp",
         "GPUGeneralizeNamedOps.cpp",
         "GPUMultiBuffering.cpp",
         "GPUPatterns.cpp",
