+++ compiler/src/iree/compiler/Codegen/Common/GPU/CMakeLists.txt
@@ -50,6 +50,7 @@ iree_cc_library(
     "GPUCheckResourceUsage.cpp"
     "GPUDistribute.cpp"
     "GPUDistributeSharedMemoryCopy.cpp"
+    "GPUDistributionPatterns.cpp"
     "GPUGeneralizeNamedOps.cpp"
     "GPUMultiBuffering.cpp"
     "GPUPatterns.cpp"
